small refactoring on backend - made socket events utils and moved start events to different file
