+++ events/SocketEvents.js
@@ -1,41 +1,28 @@
 const getCurrentTimestamp = require('../utils/utils').getCurrentTimestamp
 const { SocketEvents, GamePhases, PlayerAffilications, ErrorMessages, PlayerRole, PolicyCards, GlobalRoomName } = require('../Dictionary')
-const { isNil, includes, find, map, pick, get, forEach, mapValues, partial, partialRight } = require('lodash')
+const { isNil, includes, find, map, pick, get, mapValues, partial, partialRight } = require('lodash')
 const ClientVerificationHof = require('../utils/ClientVerificationHof')
 const RoomsManager = new (require('../utils/RoomsManager'))()
-let cancelTimeoutToken
+const SocketEventsUtils = require('../utils/SocketEventsUtils')
+const PhaseSocketEvents = require('./PhaseSocketEvents')
 
 module.exports = function (io) {
-    const facistSubproperties = ['playerName', 'affiliation', 'facistAvatar']
+    const socketEventsUtils = SocketEventsUtils(io, RoomsManager)
+    const phaseSocketEvents = PhaseSocketEvents(io, RoomsManager)
     const socketEvents = {
-        sendError: (socket, errorMessage) => {
-            socket.emit(SocketEvents.CLIENT_ERROR, { error: errorMessage })
-        },
-
-        sendBecomeFascist: (player, playerCount, passedFacists) => {
-            const shouldHideOtherFacists = player.affiliation === PlayerAffilications.HITLER_AFFILIATION && playerCount > 6
-            player.emit(SocketEvents.BECOME_FACIST, {
-                data: {
-                    facists: (shouldHideOtherFacists
-                        ? pick(player, facistSubproperties)
-                        : passedFacists
-                    ),
-                },
-            })
-        },
         checkForImmediateSuperpowersOrContinue: (socket) => {
             const fascistPolicyCount = RoomsManager.getPolicyCardsCount(socket.currentRoom, PolicyCards.FacistPolicy)
 
             // 4th power is always kill on each board
-            if (fascistPolicyCount === 4 || fascistPolicyCount === 5) {
-                socketEvents.startKillPhase(socket)
+            if (fascistPolicyCount === 4) {
+                phaseSocketEvents.startKillPhase(socket)
             // 5th power is always kill AND veto power unlock
             } else if (fascistPolicyCount === 5) {
                 RoomsManager.toggleVeto(socket.currentRoom)
-                socketEvents.sendMessage(socket, { content: 'The veto power has been unlocked! Now president or chancellor can veto any enacted policy!' })
-                socketEvents.resumeGame(socket, { delay: 3000, func: socketEvents.startChancellorChoicePhase })
+                socketEventsUtils.sendMessage(socket, { content: 'The veto power has been unlocked! Now president or chancellor can veto any enacted policy!' })
+                phaseSocketEvents.startKillPhase(socket)
             } else {
-                socketEvents.resumeGame(socket, { delay: 3000, func: socketEvents.startChancellorChoicePhase })
+                socketEventsUtils.resumeGame(socket, { delay: 3000, func: phaseSocketEvents.startChancellorChoicePhase })
             }
         },
         triggerVetoPrompt: (socket) => {
@@ -46,42 +33,38 @@ module.exports = function (io) {
             const chancellorEmit = RoomsManager.getRoleSocket(socket.currentRoom, PlayerRole.ROLE_CHANCELLOR)
             presidentEmit(SocketEvents.ServerWaitingForVeto)
             chancellorEmit(SocketEvents.ServerWaitingForVeto)
-            socketEvents.resumeGame(
+            socketEventsUtils.resumeGame(
                 socket,
                 {
                     delay: 30000,
-                    func: socketEvents.startChancellorChoicePhase,
+                    func: phaseSocketEvents.startChancellorChoicePhase,
                     customMessage: 'Due to veto power, the president and chancellor can now together veto the enacted policy. Next phase will begin in 30 seconds (assuming no veto will be reported)...',
                 },
             )
         },
-       
+
         // protect it that it can only be fired by president OR chancellor ONCE
         veto: (socket) => {
             const gamePhase = RoomsManager.getGamePhase(socket.currentRoom)
             if (gamePhase !== GamePhases.ServerWaitingForVeto) {
                 console.error('Player tried to veto when the server was not waiting for it!')
-                socket.emit(SocketEvents.CLIENT_ERROR, {
-                    error: 'You cannot veto right now!',
-                })
+                socketEventsUtils.sendError(socket, 'You cannot veto right now!')
             }
             const vetoVotes = RoomsManager.getVetoVotes(socket.currentRoom)
             const playerRole = RoomsManager.getPlayerRole(socket.currentRoom, socket.currentPlayerName)
             if (includes(vetoVotes, playerRole)) {
                 console.error('Player tried to vote twice!')
-                socket.emit(SocketEvents.CLIENT_ERROR, {
-                    error: 'You cannot veto twice!',
-                })
+                socketEventsUtils.sendError(socket, 'You cannot veto twice!')
             }
 
             RoomsManager.addVetoVote(socket.currentRoom, socket.currentPlayerName)
             const roleString = playerRole === PlayerRole.ROLE_PRESIDENT ? 'president' : 'chancellor'
             if (RoomsManager.didVetoSucceed(socket.currentRoom)) {
                 RoomsManager.setGamePhase(socket.currentRoom, GamePhases.ServerAcceptedVeto)
-                socketEvents.sendMessage(socket, { content: `The ${roleString} invoked veto for the enacted policy as well! The enacted policy has been rejected!` })
-                clearTimeout(cancelTimeoutToken)
+                socketEventsUtils.sendMessage(socket, { content: `The ${roleString} invoked veto for the enacted policy as well! The enacted policy has been rejected!` })
+                socketEventsUtils.clearNextPhaseTimeout()
                 RoomsManager.discardPolicyByVeto(socket.currentRoom)
-                socketEvents.checkIfTrackerPositionShouldUpdate(socket, false)
+                socketEventsUtils.checkIfTrackerPositionShouldUpdate(socket, false)
 
                 io.sockets.in(socket.currentRoom).emit(SocketEvents.SyncPolicies, {
                     data: {
@@ -89,21 +72,11 @@ module.exports = function (io) {
                         liberal: RoomsManager.getPolicyCardsCount(socket.currentRoom, PolicyCards.LiberalPolicy),
                     },
                 })
-                
-                socketEvents.resumeGame(socket, { delay: 5000, func: socketEvents.startChancellorChoicePhase })
+
+                socketEventsUtils.resumeGame(socket, { delay: 5000, func: phaseSocketEvents.startChancellorChoicePhase })
             } else {
                 const missingVetoRoleString = playerRole === PlayerRole.ROLE_PRESIDENT ? 'chancellor' : 'president'
-                socketEvents.sendMessage(socket, { content: `The ${roleString} invoked veto for the enacted policy! Will the ${missingVetoRoleString} call veto as well?` })
-            }
-        },
-
-        resumeGame: (socket, { delay, func, customMessage }) => {
-            const gamePhase = RoomsManager.getGamePhase(socket.currentRoom)
-            if (delay) {
-                socketEvents.sendMessage(socket, { content: customMessage || `Next phase will begin in ${delay / 1000} seconds!` })
-                cancelTimeoutToken = setTimeout(func.bind(null, socket), delay)
-            } else {
-                func(socket)
+                socketEventsUtils.sendMessage(socket, { content: `The ${roleString} invoked veto for the enacted policy! Will the ${missingVetoRoleString} call veto as well?` })
             }
         },
 
@@ -129,8 +102,7 @@ module.exports = function (io) {
                             || newOwner.affiliation === PlayerAffilications.HITLER_AFFILIATION) {
                             const playersCount = RoomsManager.getPlayersCount(socket.currentRoom)
                             const fascists = RoomsManager.getFacists(socket.currentRoom)
-                            const passedFacists = map(fascists, fascist => pick(fascist, facistSubproperties))
-                            socketEvents.sendBecomeFascist(newOwner, playersCount, passedFacists)
+                            socketEventsUtils.sendBecomeFascist(newOwner, playersCount, fascists)
                         }
                         newOwner.emit(SocketEvents.CLIENT_SEND_MESSAGE, {
                             data: {
@@ -153,7 +125,7 @@ module.exports = function (io) {
             }
 
             RoomsManager.removePlayerFromPlayersList(socket.currentPlayerName)
-            socketEvents.switchRooms(socket, socket.currentRoom, '')
+            socketEventsUtils.switchRooms(socket, socket.currentRoom, '')
             socket.currentPlayerName = ''
         },
 
@@ -177,16 +149,6 @@ module.exports = function (io) {
             }
         },
 
-        sendMessage: (socket, { content, author }) => {
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.CLIENT_SEND_MESSAGE, {
-                data: {
-                    timestamp: getCurrentTimestamp(),
-                    author,
-                    content,
-                },
-            })
-        },
-
         joinRoom: (socket, { playerName, roomName }) => {
             if (roomName && socket.currentRoom === GlobalRoomName && RoomsManager.isRoomPresent(roomName)) {
                 if (RoomsManager.isInBlackList(roomName, playerName)) {
@@ -198,7 +160,7 @@ module.exports = function (io) {
                 }
 
                 RoomsManager.addPlayer(roomName, playerName, socket)
-                socketEvents.switchRooms(socket, socket.currentRoom, roomName)
+                socketEventsUtils.switchRooms(socket, socket.currentRoom, roomName)
 
                 const roomDetails = RoomsManager.getRoomDetails(roomName)
 
@@ -213,97 +175,7 @@ module.exports = function (io) {
                 })
             } else {
                 console.error(`ERROR - Why is the room gone!, Player ${playerName} tried to enter nonexistent room ${roomName}!`)
-                socket.emit(SocketEvents.CLIENT_ERROR, {
-                    error: 'Error - WHY IS THE ROOM GONE?!',
-                })
-            }
-        },
-
-        startGame: (socket) => {
-            RoomsManager.startGame(socket.currentRoom)
-            const facists = RoomsManager.getFacists(socket.currentRoom)
-
-            // just filtering out emit functions
-            const passedFacists = map(facists, facist => pick(facist, facistSubproperties))
-            const playerCount = RoomsManager.getPlayersCount(socket.currentRoom)
-
-            forEach(facists, player => socketEvents.sendBecomeFascist(player, playerCount, passedFacists))
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.START_GAME, {
-                data: {
-                    playerName: socket.currentPlayerName,
-                    timestamp: getCurrentTimestamp(),
-                },
-            })
-        },
-
-        startVotingPhaseVote: (socket, { playerName: chancellorName }) => {
-            RoomsManager.initializeVoting(socket.currentRoom, chancellorName)
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.VOTING_PHASE_START, {
-                data: {
-                    chancellorCandidate: chancellorName,
-                    timestamp: getCurrentTimestamp(),
-                },
-            })
-        },
-
-        startChancellorChoicePhase: (socket) => {
-            RoomsManager.startChancellorChoicePhase(socket.currentRoom)
-            const playersChoices = RoomsManager.getChancellorChoices(socket.currentRoom)
-
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.CHANCELLOR_CHOICE_PHASE, {
-                data: {
-                    playersChoices,
-                    presidentName: RoomsManager.getPresident(socket.currentRoom).playerName,
-                    timestamp: getCurrentTimestamp(),
-                },
-            })
-        },
-
-        enactPolicy: (socket, policy) => {
-            const isFacist = policy === PolicyCards.FacistPolicy
-            RoomsManager.enactPolicy(socket.currentRoom, policy)
-            socketEvents.sendMessage(socket, { content: `A ${isFacist ? 'facist' : 'liberal'} policy has been enacted!` })
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.NewPolicy, {
-                data: {
-                    policy,
-                },
-            })
-        },
-
-        resetElectionTracker: (socket, positionBeforeReset) => {
-            if (positionBeforeReset) {
-                io.sockets.in(socket.currentRoom).emit(SocketEvents.ResetTracker, {
-                    data: {
-                        timestamp: getCurrentTimestamp(),
-                        trackerPositionBeforeReset: positionBeforeReset,
-                    },
-                })
-                RoomsManager.resetFailedElectionsCount(socket.currentRoom)
-                const trackerMessage = `The failed elections tracker${positionBeforeReset === 3 ? ' has reached 3, so it' : ''} will be reset!`
-                socketEvents.sendMessage(socket, { content: trackerMessage })
-            }
-        },
-
-        checkIfTrackerPositionShouldUpdate: (socket, isSuccess) => {
-            if (isSuccess) {
-                const trackerPosition = RoomsManager.getFailedElectionsCount(socket.currentRoom)
-                socketEvents.resetElectionTracker(socket, trackerPosition)
-            } else {
-                RoomsManager.increaseFailedElectionsCount(socket.currentRoom)
-                socketEvents.sendMessage(socket, { content: 'The failed elections tracker has increased!' })
-                const failedElectionsCount = RoomsManager.getFailedElectionsCount(socket.currentRoom)
-                if (failedElectionsCount >= 3) {
-                    socketEvents.resetElectionTracker(socket, failedElectionsCount)
-
-                    const topCard = RoomsManager.takeChoicePolicyCards(socket.currentRoom, 1)
-                    socketEvents.enactPolicy(socket, topCard)
-                } else {
-                    io.sockets.in(socket.currentRoom).emit(SocketEvents.IncreaseTrackerPosition, {
-                        data: {
-                            timestamp: getCurrentTimestamp(),
-                        },
-                    })
-                }
+                socketEventsUtils.sendError(socket, 'Error = WHY IS THE ROOM GONE?!')
             }
         },
 
@@ -322,15 +194,15 @@ module.exports = function (io) {
                     ? `${socket.currentRoom} has become the new chancellor!`
                     : 'The proposal has been rejected!'}
                 `
-                socketEvents.sendMessage(socket, { content: votingResultMessage })
+                socketEventsUtils.sendMessage(socket, { content: votingResultMessage })
 
-                socketEvents.checkIfTrackerPositionShouldUpdate(socket, hasVotingSucceed)
+                socketEventsUtils.checkIfTrackerPositionShouldUpdate(socket, hasVotingSucceed)
 
                 if (hasVotingSucceed) {
                     RoomsManager.setChancellor(socket.currentRoom)
-                    socketEvents.resumeGame(socket, { delay: 3000, func: socketEvents.startPresidentPolicyChoice })
+                    socketEventsUtils.resumeGame(socket, { delay: 3000, func: phaseSocketEvents.startPresidentPolicyChoice })
                 } else {
-                    socketEvents.resumeGame(socket, { delay: 3000, func: socketEvents.startChancellorChoicePhase })
+                    socketEventsUtils.resumeGame(socket, { delay: 3000, func: phaseSocketEvents.startChancellorChoicePhase })
                 }
 
                 io.sockets.in(socket.currentRoom).emit(SocketEvents.VOTING_PHASE_REVEAL, {
@@ -344,24 +216,6 @@ module.exports = function (io) {
                     },
                 })
             }
-        },
-        startPresidentPolicyChoice: (socket) => {
-            const presidentEmit = RoomsManager.getRoleSocket(socket.currentRoom, PlayerRole.ROLE_PRESIDENT)
-
-            RoomsManager.setGamePhase(socket.currentRoom, GamePhases.PresidentPolicyChoice)
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.PresidentChoosePolicy, {
-                data: {
-                    timestamp: getCurrentTimestamp(),
-                    presidentName: RoomsManager.getPresident(socket.currentRoom).playerName,
-                },
-            })
-            presidentEmit(SocketEvents.ChoosePolicy, {
-                data: {
-                    policyCards: RoomsManager.takeChoicePolicyCards(socket.currentRoom, 3),
-                    title: 'Discard one policy and pass the rest to the chancellor',
-                    role: PlayerRole.ROLE_PRESIDENT,
-                },
-            })
         },
 
         choosePolicy: (socket, { choice }) => {
@@ -391,7 +245,7 @@ module.exports = function (io) {
                     RoomsManager.discardPolicy(socket.currentRoom, choice)
                     const enactedPolicy = drawnCards[0]
 
-                    socketEvents.enactPolicy(socket, enactedPolicy)
+                    socketEventsUtils.enactPolicy(socket, enactedPolicy)
 
                     const isVetoUnlocked = RoomsManager.isVetoUnlocked(socket.currentRoom)
                     if (isVetoUnlocked) {
@@ -399,7 +253,7 @@ module.exports = function (io) {
                     } else if (enactedPolicy === PolicyCards.FacistPolicy) {
                         socketEvents.checkForImmediateSuperpowersOrContinue(socket)
                     } else {
-                        socketEvents.resumeGame(socket, { delay: 3000, func: socketEvents.startChancellorChoicePhase })
+                        socketEventsUtils.resumeGame(socket, { delay: 3000, func: phaseSocketEvents.startChancellorChoicePhase })
                     }
                 } else {
                     console.error('Cheater!')
@@ -409,19 +263,6 @@ module.exports = function (io) {
                 console.error('Cheater!')
                 // cheating (sent card was not in collection)! 
             }
-        },
-
-        startKillPhase: (socket) => {
-            RoomsManager.setGamePhase(socket.currentRoom, GamePhases.GAME_PHASE_SUPERPOWER)
-            const presidentName = get(RoomsManager.getPresident(socket.currentRoom), 'playerName')
-            const playersChoices = RoomsManager.getOtherAlivePlayers(socket.currentRoom, presidentName)
-            io.sockets.in(socket.currentRoom).emit(SocketEvents.KillSuperpowerUsed, {
-                data: {
-                    presidentName,
-                    timestamp: getCurrentTimestamp(),
-                    playersChoices,
-                },
-            })
         },
 
         killPlayer: (socket, { playerName }) => {
@@ -447,7 +288,7 @@ module.exports = function (io) {
                     },
                 })
             } else {
-                socketEvents.resumeGame(socket, { delay: 4000, func: socketEvents.startChancellorChoicePhase })
+                socketEventsUtils.resumeGame(socket, { delay: 4000, func: phaseSocketEvents.startChancellorChoicePhase })
             }
         },
         kickPlayer: (socket, { playerName }, permanently = false) => {
@@ -460,52 +301,13 @@ module.exports = function (io) {
                 },
             })
             const kickedSocket = find(io.sockets.in(socket.currentRoom).sockets, { currentPlayerName: playerName })
-            socketEvents.switchRooms(kickedSocket, socket.currentRoom, GlobalRoomName)
-        },
-        switchRooms: (socket, startRoom, targetRoom) => {
-            if (startRoom) {
-                socket.leave(startRoom)
-                socketEvents.sendMessage(socket, { content: `${socket.currentPlayerName} has left the room` })
-                
-                const updatedRoom = (startRoom === GlobalRoomName ? targetRoom : startRoom)
-                if (updatedRoom) {
-                    io.sockets.in(GlobalRoomName).emit(SocketEvents.RoomsListChanged, {
-                        data: {
-                            roomName: updatedRoom,
-                            room: RoomsManager.getRoomDetailsForLobby(updatedRoom),
-                        },
-                    })
-                }
-            }
-            socket.currentRoom = targetRoom
-            if (targetRoom) {
-                if (startRoom) {
-                    RoomsManager.updatePlayerRoom(socket.currentPlayerName, targetRoom)
-                }
-
-                if (targetRoom === GlobalRoomName) {
-                    socket.emit(SocketEvents.SyncLobby, {
-                        data: {
-                            players: RoomsManager.getPlayersList(),
-                            rooms: RoomsManager.getRoomsList(),
-                        },
-                    })
-                }
-                socket.join(targetRoom)
-                socketEvents.sendMessage(socket, { content: `${socket.currentPlayerName} has joined the room` })
-            }
-            io.sockets.in(GlobalRoomName).emit(SocketEvents.PlayersListChanged, {
-                data: {
-                    playerName: socket.currentPlayerName,
-                    player: RoomsManager.getPlayerFromPlayersList(socket.currentPlayerName),
-                },
-            })
+            socketEventsUtils.switchRooms(kickedSocket, socket.currentRoom, GlobalRoomName)
         },
         selectName: (socket, { userName }) => {
             // deselecting name
             if (!userName) {
                 RoomsManager.removePlayerFromPlayersList(socket.currentPlayerName)
-                socketEvents.switchRooms(socket, GlobalRoomName, '')
+                socketEventsUtils.switchRooms(socket, GlobalRoomName, '')
                 socket.emit(SocketEvents.SelectName, { data: { userName: '' } })
                 socket.currentPlayerName = ''
             // selecting name
@@ -513,9 +315,9 @@ module.exports = function (io) {
                 RoomsManager.addPlayerToPlayersList(userName)
                 socket.emit(SocketEvents.SelectName, { data: { userName } })
                 socket.currentPlayerName = userName
-                socketEvents.switchRooms(socket, '', GlobalRoomName)
+                socketEventsUtils.switchRooms(socket, '', GlobalRoomName)
             } else {
-                socketEvents.sendError(socket, ErrorMessages.NameTaken)
+                socketEventsUtils.sendError(socket, ErrorMessages.NameTaken)
             }
         },
     }
@@ -525,12 +327,16 @@ module.exports = function (io) {
         socket.currentRoom = ''
 
         const clientVerificationHof = ClientVerificationHof(RoomsManager)
-        socketEvents.startGame = clientVerificationHof(['isOwner'], socketEvents.startGame)
+        socketEvents.startGame = clientVerificationHof(['isOwner'], phaseSocketEvents.startGame)
         socketEvents.kickPlayer = clientVerificationHof(['isOwner'], socketEvents.kickPlayer)
 
         // to avoid creating new binded functions each time an action is made. This is made only once.
         // we need a way to pass socket object into those functions
-        const partialFunctions = mapValues(socketEvents, func => partial(func, socket))
+        const partialFunctions = mapValues({
+            ...socketEvents,
+            ...phaseSocketEvents,
+            sendMessage: socketEventsUtils.sendMessage,
+        }, func => partial(func, socket))
         partialFunctions['banPlayer'] = partialRight(partialFunctions.kickPlayer, true)
 
         socket.on('disconnect', partialFunctions.disconnect)
@@ -546,6 +352,6 @@ module.exports = function (io) {
         socket.on(SocketEvents.PlayerKicked, partialFunctions.kickPlayer)
         socket.on(SocketEvents.ChoosePolicy, partialFunctions.choosePolicy)
         socket.on(SocketEvents.SelectName, partialFunctions.selectName)
-        socket.on(SocketEvents.VetoRegistered, partialFunctions.veto)
+        socket.on(SocketEvents.VetoVoteRegistered, partialFunctions.veto)
     })
 }
